pyproject.toml
[tool.poetry]
name = "saki-chan"
version = "0.1.0"
description = ""
authors = ["Giang Ta <thoanggiang98@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^1.0.1"
python-dotenv = "^0.15.0"
riotwatcher = "^3.1.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pylint = "^2.6.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
